# LUCENE-9932: Performance improvement for BKD index building