--- a/lucene/core/src/java/org/apache/lucene/index/PointValuesWriter.java
+++ b/lucene/core/src/java/org/apache/lucene/index/PointValuesWriter.java
@@ -313,5 +313,14 @@ class PointValuesWriter {
     public void restore(int i, int j) {
       in.restore(i, j);
     }
+
+    @Override
+    public boolean indexSortingReorder(int i, int j) {
+      for (int k = i; k < j; k++) {
+        save(in.getDocID(i), docMap.oldToNew(in.getDocID(i)));
+      }
+      restore(i, j);
+      return true;
+    }
   }
 }