--- a/lucene/core/src/java/org/apache/lucene/index/PointValuesWriter.java
+++ b/lucene/core/src/java/org/apache/lucene/index/PointValuesWriter.java
@@ -313,6 +313,5 @@ class PointValuesWriter {
     public void restore(int i, int j) {
       in.restore(i, j);
     }
-
   }
 }