--- a/lucene/core/src/java/org/apache/lucene/util/MSBRadixSorter.java
+++ b/lucene/core/src/java/org/apache/lucene/util/MSBRadixSorter.java
@@ -31,18 +31,18 @@ public abstract class MSBRadixSorter extends Sorter {
   // this is used as a protection against the fact that radix sort performs
   // worse when there are long common prefixes (probably because of cache
   // locality)
-  private static final int LEVEL_THRESHOLD = 8;
+  protected static final int LEVEL_THRESHOLD = 8;
   // size of histograms: 256 + 1 to indicate that the string is finished
-  private static final int HISTOGRAM_SIZE = 257;
+  protected static final int HISTOGRAM_SIZE = 257;
   // buckets below this size will be sorted with introsort
-  private static final int LENGTH_THRESHOLD = 100;
+  protected static final int LENGTH_THRESHOLD = 100;
 
   // we store one histogram per recursion level
-  private final int[][] histograms = new int[LEVEL_THRESHOLD][];
-  private final int[] endOffsets = new int[HISTOGRAM_SIZE];
-  private final int[] commonPrefix;
+  protected final int[][] histograms = new int[LEVEL_THRESHOLD][];
+  protected final int[] endOffsets = new int[HISTOGRAM_SIZE];
+  protected int[] commonPrefix;
 
-  private final int maxLength;
+  protected int maxLength;
 
   /**
    * Sole constructor.
@@ -128,7 +128,7 @@ public abstract class MSBRadixSorter extends Sorter {
     sort(from, to, 0, 0);
   }
 
-  private void sort(int from, int to, int k, int l) {
+  protected void sort(int from, int to, int k, int l) {
     if (to - from <= LENGTH_THRESHOLD || l >= LEVEL_THRESHOLD) {
       introSort(from, to, k);
     } else {
@@ -202,7 +202,7 @@ public abstract class MSBRadixSorter extends Sorter {
   }
 
   /** Return a number for the k-th character between 0 and {@link #HISTOGRAM_SIZE}. */
-  private int getBucket(int i, int k) {
+  protected int getBucket(int i, int k) {
     return byteAt(i, k) + 1;
   }
 
@@ -284,7 +284,7 @@ public abstract class MSBRadixSorter extends Sorter {
    * @param startOffsets start offsets per bucket
    * @param endOffsets end offsets per bucket
    */
-  private void reorder(int from, int to, int[] startOffsets, int[] endOffsets, int k) {
+  protected void reorder(int from, int to, int[] startOffsets, int[] endOffsets, int k) {
     // reorder in place, like the dutch flag problem
     for (int i = 0; i < HISTOGRAM_SIZE; ++i) {
       final int limit = endOffsets[i];