--- a/lucene/core/src/java/org/apache/lucene/util/bkd/MutablePointsReaderUtils.java
+++ b/lucene/core/src/java/org/apache/lucene/util/bkd/MutablePointsReaderUtils.java
@@ -38,12 +38,15 @@ public final class MutablePointsReaderUtils {
   MutablePointsReaderUtils() {}
 
   /**
-   * Sort the given {@link MutablePointValues} based on its packed value, note that doc ID is not
-   * taken into sorting algorithm, since if they are already in ascending order, stable sort is able
-   * to maintain the ordering of doc ID.
+   * Sort the given {@link MutablePointValues} based on its packed value. If doc IDs are already
+   * in ascending order, stable sort is able to maintain the ordering of doc ID. If index sorting
+   * is enabled, we can leverage sortMap to reorder without having to sort (ie. O(n) rather than O(n log n)).
    */
   public static void sort(
       BKDConfig config, int maxDoc, MutablePointValues reader, int from, int to) {
+    if (reader.indexSortingReorder(from, to)) {
+      return;
+    }
     new StableMSBRadixSorter(config.packedBytesLength) {
 
       @Override
@@ -68,31 +71,6 @@ public final class MutablePointsReaderUtils {
               "k should be less than packedBytesLength " + config.packedBytesLength);
         }
         return Byte.toUnsignedInt(reader.getByteAt(i, k));
-      }
-
-      @Override
-      protected Sorter getFallbackSorter(int k) {
-        return new InPlaceMergeSorter() {
-
-          @Override
-          protected int compare(final int i, final int j) {
-            for (int o = k; o < config.packedBytesLength; ++o) {
-              final int b1 = byteAt(i, o);
-              final int b2 = byteAt(j, o);
-              if (b1 != b2) {
-                return b1 - b2;
-              } else if (b1 == -1) {
-                break;
-              }
-            }
-            return 0;
-          }
-
-          @Override
-          protected void swap(final int i, final int j) {
-            reader.swap(i, j);
-          }
-        };
       }
     }.sort(from, to);
   }