… lucene/CHANGES.txt
… lucene/core/src/test/org/apache/lucene/util/TestStableMSBRadixSorter.java
… lucene/core/src/test/org/apache/lucene/util/bkd/TestBKD.java
… lucene/core/src/test/org/apache/lucene/util/bkd/TestBKDDisableSortDocId.java
… lucene/core/src/test/org/apache/lucene/util/bkd/TestMutablePointsReaderUtils.java