@jpountz 2021-04-19
Wow, this is a neat idea! I wonder if we could find a less intrusive way to apply this change, e.g. could we avoid adding info about whether to sort by doc ID to BKDConfig? For instance could MSBRadixSorter detect whether data is sorted by itself instead of having to propagate this information in so many places?