@neoremind 2021-04-20
@jpountz Good advice! Before that I am still struggling where to propagate this config up to the index builder layer. 
I will give it a try, the first thing comes up my mind is to bring up a new `prepare` method, in which it will scan all docid from i to j to see if they are increasing. I will experiment on this, if the overhead is small enough, then it is worthwhile to sort without docid.
One more question, are there any places where doc id is not added increasingly? I mean the source code, not test cases.