@jpountz 2021-04-20
> I will give it a try, the first thing comes up my mind is to bring up a new prepare method,

One idea I had in mind was to create a new class, something like `StableMSBRadixSorter` that would extend `MSBRadixSorter` to:
 - add the two `assign` and `finalizeAssign` methods that you currently added to `Sorter`,
 - override the way data gets rearranged to guarantee stability,
 - change the fallback sorter,
 - modify `radixSort(int,int,int,int)` to check whether data is already sorted before computing the common prefix length and the histogram of the leading bytes.

> One more question, are there any places where doc id is not added increasingly?

I don't remember how we deal with it, but we should check how this optimization plays with index sorting, since we would renumber doc IDs at flush time.