@jpountz 2021-04-21
I left a few comments before your previous comment, I left some suggestions to make the code simpler but I think that it's getting close.

Thanks for the great performance analysis! Given that you are benchmarking with fully random bytes, I'm actually expecting the speedup to be greater with real data where the leading bytes are often zeroes. For instance I'd expect users who index integers (4 bytes) between 0 and 2^24 to notice speedups that are closer to the one that you computed for bytesPerDim=3 than for bytesPerDim=4.