@iverase 2021-04-21
What I read for this results is that the stable sorting is faster than the non-stable. I think it makes sense because re-ordering using a helper array is probably faster that in-place re-ordering. 

If we are ok to take the memory penalty of using this extra array, then I think it make sense to always use the stable version of the algorithm?