@neoremind 2021-04-21
> For instance I'd expect users who index integers (4 bytes) between 0 and 2^24 to notice speedups that are closer to the one that you computed for bytesPerDim=3 than for bytesPerDim=4.
I can also construct a real case to verify :smile: