@jpountz 2021-04-21
> If we are ok to take the memory penalty of using this extra array, then I think it make sense to always use the stable version of the algorithm?

+1 to always use the stable version of the algorithm. This would only use transient memory and in reasonable amounts, so I'm not concerned with the memory usage.